Move app.js to app
